annotate part of the functions so that it can run

+++ appointment_system/src/main/java/com/cpt202/appointment_system/Repositories/UserRepo.java
@@ -9,11 +9,11 @@ import com.cpt202.appointment_system.Models.User;
 
 public interface UserRepo extends JpaRepository<User, Integer> {
 
-    public List<User> findByType(byte type);
+    public List<User> findByType(Integer type);
 
     public List<User> findByUsernameContaining(String username);
 
-    public User findByUid(int uid);
+    public User findByUid(Integer uid);
 
     public boolean existsByUsername(String username);
